🌱 refactor: split high-complexity components into sub-components
Extract sub-components from 5 large files to improve maintainability:

- NamespaceManager.tsx (1421→703 lines): Extract modals and card components
- PodDrillDown.tsx (2398→1550 lines): Extract tab content components
- GPUReservations.tsx (1457→698 lines): Extract tab and chart components
- AddClusterDialog.tsx (914→349 lines): Extract tab components
- Clusters.tsx (811→435 lines): Extract custom hooks and groups section

Closes #3465
