+++ web/src/components/clusters/Clusters.tsx
@@ -1,6 +1,6 @@
 import { useState, useMemo, useEffect, useCallback } from 'react'
 import { useSearchParams, useLocation } from 'react-router-dom'
-import { Check, WifiOff, ChevronRight, CheckCircle, AlertTriangle, ChevronDown, FolderOpen, Plus, Trash2, Server, Scissors } from 'lucide-react'
+import { AlertTriangle, ChevronRight, ChevronDown, Server, Scissors } from 'lucide-react'
 import { useClusters, useGPUNodes, useNVIDIAOperators, refreshSingleCluster } from '../../hooks/useMCP'
 import { ClusterDetailModal } from './ClusterDetailModal'
 import { AddClusterDialog } from './AddClusterDialog'
@@ -12,8 +12,6 @@ import {
   GPUDetailModal,
   type ClusterLayoutMode,
 } from './components'
-import { isClusterUnreachable, isClusterHealthy } from './utils'
-import { detectCloudProvider, getProviderLabel } from '../ui/CloudProviderIcon'
 import { useMissions } from '../../hooks/useMissions'
 import { useApiKeyCheck, ApiKeyPromptModal } from '../cards/console-missions/shared'
 import { loadMissionPrompt } from '../cards/multi-tenancy/missionLoader'
@@ -35,6 +33,9 @@ import { useUniversalStats, createMergedStatValueGetter } from '../../hooks/useU
 import type { StatBlockValue } from '../ui/StatsOverview'
 import { formatMemoryStat } from '../../lib/formatStats'
 import { RotatingTip } from '../ui/RotatingTip'
+import { useClusterFiltering } from './useClusterFiltering'
+import { useClusterStats } from './useClusterStats'
+import { ClusterGroupsSection } from './ClusterGroupsSection'
 
 // Storage key for cluster page cards
 const CLUSTERS_CARDS_KEY = 'kubestellar-clusters-cards'
@@ -113,10 +114,6 @@ export function Clusters() {
     return (stored as ClusterLayoutMode) || 'grid'
   })
   const [renamingCluster, setRenamingCluster] = useState<string | null>(null)
-  const [showGroupForm, setShowGroupForm] = useState(false)
-  const [newGroupName, setNewGroupName] = useState('')
-  const [newGroupClusters, setNewGroupClusters] = useState<string[]>([])
-  const [showGroups, setShowGroups] = useState(false) // Collapsed by default so cluster cards are visible first
 
   // Additional UI state
   const [showClusterGrid, setShowClusterGrid] = useState(true) // Cluster cards visible by default
@@ -149,79 +146,16 @@ export function Clusters() {
     safeSetItem(STORAGE_KEY_CLUSTER_ORDER, JSON.stringify(newOrder))
   }, [])
 
-  const filteredClusters = useMemo(() => {
-    let result = clusters || []
-
-    // Apply global cluster filter
-    if (!isAllClustersSelected) {
-      result = result.filter(c => globalSelectedClusters.includes(c.name))
-    }
-
-    // Apply custom text filter
-    if (customFilter.trim()) {
-      const query = customFilter.toLowerCase()
-      result = result.filter(c =>
-        c.name.toLowerCase().includes(query) ||
-        c.context?.toLowerCase().includes(query) ||
-        c.server?.toLowerCase().includes(query) ||
-        c.user?.toLowerCase().includes(query)
-      )
-    }
-
-    // Apply local health filter
-    if (filter === 'healthy') {
-      result = result.filter(c => !isClusterUnreachable(c) && isClusterHealthy(c))
-    } else if (filter === 'unhealthy') {
-      result = result.filter(c => !isClusterUnreachable(c) && !isClusterHealthy(c))
-    } else if (filter === 'unreachable') {
-      result = result.filter(c => isClusterUnreachable(c))
-    }
-
-    // Sort
-    if (sortBy === 'custom' && customOrder.length > 0) {
-      // Use custom order: items in customOrder come first in that order,
-      // items not in customOrder are appended alphabetically
-      const orderMap = new Map(customOrder.map((name, i) => [name, i]))
-      result = [...result].sort((a, b) => {
-        const aIdx = orderMap.get(a.name)
-        const bIdx = orderMap.get(b.name)
-        if (aIdx !== undefined && bIdx !== undefined) return aIdx - bIdx
-        if (aIdx !== undefined) return -1
-        if (bIdx !== undefined) return 1
-        return a.name.localeCompare(b.name)
-      })
-    } else {
-      result = [...result].sort((a, b) => {
-        let cmp = 0
-        switch (sortBy) {
-          case 'name':
-            cmp = a.name.localeCompare(b.name)
-            break
-          case 'nodes':
-            cmp = (a.nodeCount || 0) - (b.nodeCount || 0)
-            break
-          case 'pods':
-            cmp = (a.podCount || 0) - (b.podCount || 0)
-            break
-          case 'health': {
-            const aHealth = isClusterUnreachable(a) ? 0 : isClusterHealthy(a) ? 2 : 1
-            const bHealth = isClusterUnreachable(b) ? 0 : isClusterHealthy(b) ? 2 : 1
-            cmp = aHealth - bHealth
-            break
-          }
-          case 'provider': {
-            const aProvider = getProviderLabel((a.distribution as ReturnType<typeof detectCloudProvider>) || detectCloudProvider(a.name, a.server, a.namespaces, a.user))
-            const bProvider = getProviderLabel((b.distribution as ReturnType<typeof detectCloudProvider>) || detectCloudProvider(b.name, b.server, b.namespaces, b.user))
-            cmp = aProvider.localeCompare(bProvider)
-            break
-          }
-        }
-        return sortAsc ? cmp : -cmp
-      })
-    }
-
-    return result
-  }, [clusters, filter, globalSelectedClusters, isAllClustersSelected, customFilter, sortBy, sortAsc, customOrder])
+  const { filteredClusters, globalFilteredClusters } = useClusterFiltering({
+    clusters,
+    filter,
+    globalSelectedClusters,
+    isAllClustersSelected,
+    customFilter,
+    sortBy,
+    sortAsc,
+    customOrder,
+  })
 
   // Get GPU count per cluster
   const gpuByCluster = useMemo(() => {
@@ -237,76 +171,7 @@ export function Clusters() {
     return map
   }, [gpuNodes])
 
-  // Base clusters after global filter (before local health filter)
-  const globalFilteredClusters = useMemo(() => {
-    let result = clusters || []
-
-    // Apply global cluster filter
-    if (!isAllClustersSelected) {
-      result = result.filter(c => globalSelectedClusters.includes(c.name))
-    }
-
-    // Apply custom text filter
-    if (customFilter.trim()) {
-      const query = customFilter.toLowerCase()
-      result = result.filter(c =>
-        c.name.toLowerCase().includes(query) ||
-        c.context?.toLowerCase().includes(query) ||
-        c.server?.toLowerCase().includes(query) ||
-        c.user?.toLowerCase().includes(query)
-      )
-    }
-
-    return result
-  }, [clusters, globalSelectedClusters, isAllClustersSelected, customFilter])
-
-  const stats = useMemo(() => {
-    // Calculate total GPUs from GPU nodes that match filtered clusters
-    // Only include GPUs from reachable clusters
-    let totalGPUs = 0
-    let allocatedGPUs = 0
-    globalFilteredClusters.forEach(cluster => {
-      // Skip offline clusters - don't count their GPUs
-      if (isClusterUnreachable(cluster)) return
-
-      const clusterKey = cluster.name.split('/')[0]
-      const gpuInfo = gpuByCluster[clusterKey] || gpuByCluster[cluster.name]
-      if (gpuInfo) {
-        totalGPUs += gpuInfo.total
-        allocatedGPUs += gpuInfo.allocated
-      }
-    })
-
-    // Separate unreachable, healthy, unhealthy - simplified logic matching sidebar
-    const unreachable = globalFilteredClusters.filter(c => isClusterUnreachable(c)).length
-    const staleContexts = globalFilteredClusters.filter(c => (c as unknown as Record<string, unknown>).neverConnected === true).length
-    const healthy = globalFilteredClusters.filter(c => !isClusterUnreachable(c) && isClusterHealthy(c)).length
-    const unhealthy = globalFilteredClusters.filter(c => !isClusterUnreachable(c) && !isClusterHealthy(c)).length
-    const loadingCount = globalFilteredClusters.filter(c =>
-      c.nodeCount === undefined && c.reachable === undefined
-    ).length
-
-    const hasResourceData = globalFilteredClusters.some(c =>
-      !isClusterUnreachable(c) && c.nodeCount !== undefined && c.nodeCount > 0
-    )
-
-    return {
-      total: globalFilteredClusters.length,
-      loading: loadingCount,
-      healthy,
-      unhealthy,
-      unreachable,
-      staleContexts,
-      totalNodes: globalFilteredClusters.reduce((sum, c) => sum + (c.nodeCount || 0), 0),
-      totalCPUs: globalFilteredClusters.reduce((sum, c) => sum + (c.cpuCores || 0), 0),
-      totalMemoryGB: globalFilteredClusters.reduce((sum, c) => sum + (c.memoryGB || 0), 0),
-      totalStorageGB: globalFilteredClusters.reduce((sum, c) => sum + (c.storageGB || 0), 0),
-      totalPods: globalFilteredClusters.reduce((sum, c) => sum + (c.podCount || 0), 0),
-      totalGPUs,
-      allocatedGPUs,
-      hasResourceData,
-    }
-  }, [globalFilteredClusters, gpuByCluster])
+  const stats = useClusterStats({ globalFilteredClusters, gpuByCluster })
 
   // Determine if we should show skeleton content (loading with no data OR offline without demo OR mode switching)
   const showSkeletonContent = (isLoading && (clusters || []).length === 0) || forceSkeletonForOffline || isModeSwitching
@@ -494,139 +359,13 @@ export function Clusters() {
       </div>
 
       {/* Cluster Groups */}
-      {(clusterGroups.length > 0 || showGroupForm) && (
-        <div className="mb-6">
-          <div className="flex items-center justify-between mb-3">
-            <button
-              onClick={() => setShowGroups(!showGroups)}
-              className="flex items-center gap-2 text-sm font-medium text-muted-foreground hover:text-foreground transition-colors"
-            >
-              <FolderOpen className="w-4 h-4" />
-              <span>Cluster Groups ({clusterGroups.length})</span>
-              {showGroups ? <ChevronDown className="w-4 h-4" /> : <ChevronRight className="w-4 h-4" />}
-            </button>
-            <button
-              onClick={() => setShowGroupForm(!showGroupForm)}
-              className="flex items-center gap-1 text-xs text-primary hover:text-primary/80 transition-colors"
-            >
-              <Plus className="w-3.5 h-3.5" />
-              New Group
-            </button>
-          </div>
-
-          {showGroups && (
-            <div className="space-y-2">
-              {/* New Group Form */}
-              {showGroupForm && (
-                <div className="glass p-4 rounded-lg space-y-3">
-                  <input
-                    type="text"
-                    placeholder="Group name..."
-                    value={newGroupName}
-                    onChange={(e) => setNewGroupName(e.target.value)}
-                    className="w-full px-3 py-2 text-sm bg-secondary/50 border border-border rounded-lg text-foreground placeholder:text-muted-foreground focus:outline-none focus:ring-1 focus:ring-primary"
-                    autoFocus
-                  />
-                  <div className="text-xs text-muted-foreground mb-1">Select clusters for this group:</div>
-                  <div className="flex flex-wrap gap-2">
-                    {clusters.map((cluster) => {
-                      const isInGroup = newGroupClusters.includes(cluster.name)
-                      const unreachable = isClusterUnreachable(cluster)
-                      return (
-                        <button
-                          key={cluster.name}
-                          onClick={() => {
-                            if (isInGroup) {
-                              setNewGroupClusters(prev => prev.filter(c => c !== cluster.name))
-                            } else {
-                              setNewGroupClusters(prev => [...prev, cluster.name])
-                            }
-                          }}
-                          className={`flex items-center gap-1.5 px-2 py-1 rounded text-xs transition-colors ${
-                            isInGroup
-                              ? 'bg-primary/20 text-primary border border-primary/30'
-                              : 'bg-secondary/50 text-muted-foreground hover:text-foreground border border-transparent'
-                          }`}
-                        >
-                          {unreachable ? (
-                            <WifiOff className="w-3 h-3 text-yellow-400" />
-                          ) : cluster.healthy ? (
-                            <CheckCircle className="w-3 h-3 text-green-400" />
-                          ) : (
-                            <AlertTriangle className="w-3 h-3 text-orange-400" />
-                          )}
-                          {cluster.context || cluster.name}
-                        </button>
-                      )
-                    })}
-                  </div>
-                  <div className="flex gap-2 justify-end">
-                    <button
-                      onClick={() => {
-                        setShowGroupForm(false)
-                        setNewGroupName('')
-                        setNewGroupClusters([])
-                      }}
-                      className="px-3 py-1.5 text-sm text-muted-foreground hover:text-foreground"
-                    >
-                      Cancel
-                    </button>
-                    <button
-                      onClick={() => {
-                        if (newGroupName.trim() && newGroupClusters.length > 0) {
-                          addClusterGroup({ name: newGroupName.trim(), clusters: newGroupClusters })
-                          setShowGroupForm(false)
-                          setNewGroupName('')
-                          setNewGroupClusters([])
-                        }
-                      }}
-                      disabled={!newGroupName.trim() || newGroupClusters.length === 0}
-                      className="flex items-center gap-1 px-3 py-1.5 text-sm bg-primary text-primary-foreground rounded-lg hover:bg-primary/80 disabled:opacity-50 disabled:cursor-not-allowed"
-                    >
-                      <Check className="w-3.5 h-3.5" />
-                      Create
-                    </button>
-                  </div>
-                </div>
-              )}
-
-              {/* Existing Groups */}
-              {clusterGroups.map((group) => (
-                <div
-                  key={group.id}
-                  className="glass p-3 rounded-lg flex items-center justify-between hover:bg-secondary/30 transition-colors"
-                >
-                  <button
-                    onClick={() => selectClusterGroup(group.id)}
-                    className="flex-1 flex items-center gap-3 text-left"
-                  >
-                    <FolderOpen className="w-4 h-4 text-purple-400" />
-                    <div>
-                      <div className="font-medium text-foreground">{group.name}</div>
-                      <div className="text-xs text-muted-foreground">
-                        {group.clusters.length} cluster{group.clusters.length !== 1 ? 's' : ''}
-                        <span className="mx-1">·</span>
-                        {group.clusters.slice(0, 3).join(', ')}
-                        {group.clusters.length > 3 && ` +${group.clusters.length - 3} more`}
-                      </div>
-                    </div>
-                  </button>
-                  <button
-                    onClick={(e) => {
-                      e.stopPropagation()
-                      deleteClusterGroup(group.id)
-                    }}
-                    className="p-1.5 rounded hover:bg-red-500/20 text-muted-foreground hover:text-red-400 transition-colors"
-                    title={t('cluster.deleteGroup')}
-                  >
-                    <Trash2 className="w-4 h-4" />
-                  </button>
-                </div>
-              ))}
-            </div>
-          )}
-        </div>
-      )}
+      <ClusterGroupsSection
+        clusters={clusters}
+        clusterGroups={clusterGroups}
+        addClusterGroup={addClusterGroup}
+        deleteClusterGroup={deleteClusterGroup}
+        selectClusterGroup={selectClusterGroup}
+      />
     </>
   )
