+++ web/src/components/clusters/add-cluster/index.ts
@@ -1,0 +1,1 @@
+export { AddClusterDialog } from '../AddClusterDialog'
