+++ web/src/components/drilldown/views/pod-drilldown/index.ts
@@ -3,3 +3,13 @@ export { UNHEALTHY_STATUSES } from './types'
 export { getIssueSeverity } from './helpers'
 export type { PodCacheEntry } from './podCache'
 export { RAPID_REOPEN_THRESHOLD_MS, getPodCache, setPodCache, cleanupPodCache } from './podCache'
+export { PodLabelsTab } from './PodLabelsTab'
+export type { PodLabelsTabProps } from './PodLabelsTab'
+export { PodRelatedTab } from './PodRelatedTab'
+export type { PodRelatedTabProps } from './PodRelatedTab'
+export { PodOutputTab } from './PodOutputTab'
+export type { PodOutputTabProps } from './PodOutputTab'
+export { PodAiAnalysis } from './PodAiAnalysis'
+export type { PodAiAnalysisProps } from './PodAiAnalysis'
+export { PodDeleteSection } from './PodDeleteSection'
+export type { PodDeleteSectionProps } from './PodDeleteSection'
