+++ web/src/components/namespaces/NamespaceManager.tsx
@@ -5,21 +5,21 @@ import {
   RefreshCw,
   Search,
   Shield,
-  UserPlus,
   Trash2,
-  ChevronRight,
   ChevronDown,
+  ChevronRight,
   X,
   AlertTriangle,
   Hourglass,
   Layers,
   Server,
+  UserPlus,
   WifiOff
 } from 'lucide-react'
 import { Button } from '../ui/Button'
 import { useClusters } from '../../hooks/useMCP'
 import { useRefreshIndicator } from '../../hooks/useRefreshIndicator'
-import { BaseModal, useModalState } from '../../lib/modals'
+import { useModalState } from '../../lib/modals'
 import { useGlobalFilters } from '../../hooks/useGlobalFilters'
 import { ClusterBadge } from '../ui/ClusterBadge'
 import { DashboardHeader } from '../shared/DashboardHeader'
@@ -28,25 +28,13 @@ import { useToast } from '../ui/Toast'
 import { useTranslation } from 'react-i18next'
 import { LOCAL_AGENT_HTTP_URL } from '../../lib/constants'
 import { NAMESPACE_ABORT_TIMEOUT_MS } from '../../lib/constants/network'
+import { NamespaceCard, NamespaceCardSkeleton } from './NamespaceCard'
+import { DeleteConfirmModal } from './DeleteConfirmModal'
+import { CreateNamespaceModal } from './CreateNamespaceModal'
+import { GrantAccessModal } from './GrantAccessModal'
+import type { NamespaceDetails, NamespaceAccessEntry } from './types'
 
 type GroupByMode = 'cluster' | 'type'
-
-interface NamespaceDetails {
-  name: string
-  cluster: string
-  status: string
-  labels?: Record<string, string>
-  createdAt: string
-}
-
-interface NamespaceAccessEntry {
-  bindingName: string
-  subjectKind: string
-  subjectName: string
-  subjectNamespace?: string
-  roleName: string
-  roleKind: string
-}
 
 // Cache for namespace data per cluster - persists across filter changes
 const namespaceCache = new Map<string, NamespaceDetails[]>()
@@ -713,709 +701,3 @@ export function NamespaceManager() {
   )
 }
 
-interface NamespaceCardProps {
-  namespace: NamespaceDetails
-  isSelected: boolean
-  onSelect: () => void
-  onDelete?: () => void
-  isSystem?: boolean
-  showCluster?: boolean
-}
-
-function NamespaceCard({ namespace, isSelected, onSelect, onDelete, isSystem, showCluster = true }: NamespaceCardProps) {
-  return (
-    <div
-      onClick={onSelect}
-      className={`flex items-center gap-4 p-4 rounded-lg cursor-pointer transition-colors group ${
-        isSelected
-          ? 'bg-blue-500/20 border border-blue-500/50'
-          : 'bg-secondary/30 hover:bg-secondary/50 border border-transparent'
-      }`}
-    >
-      <div className={`w-10 h-10 rounded-lg flex items-center justify-center ${
-        isSystem ? 'bg-gray-500/20' : 'bg-blue-500/20'
-      }`}>
-        <Folder className={`w-5 h-5 ${isSystem ? 'text-muted-foreground' : 'text-blue-400'}`} />
-      </div>
-      <div className="flex-1">
-        <div className="flex items-center gap-2">
-          <span className="font-medium text-white">{namespace.name}</span>
-          <span className={`text-xs px-1.5 py-0.5 rounded ${
-            namespace.status === 'Active'
-              ? 'bg-green-500/20 text-green-400'
-              : 'bg-yellow-500/20 text-yellow-400'
-          }`}>
-            {namespace.status}
-          </span>
-        </div>
-        <p className="text-sm text-muted-foreground">
-          Created {new Date(namespace.createdAt).toLocaleDateString()}
-        </p>
-      </div>
-      {showCluster && <ClusterBadge cluster={namespace.cluster} size="sm" />}
-      {!isSystem && onDelete && (
-        <button
-          onClick={(e) => {
-            e.stopPropagation()
-            onDelete()
-          }}
-          className="p-2 rounded text-muted-foreground hover:text-red-400 hover:bg-red-500/10 transition-colors opacity-0 group-hover:opacity-100"
-          title="Delete namespace"
-        >
-          <Trash2 className="w-4 h-4" />
-        </button>
-      )}
-      <ChevronRight className="w-4 h-4 text-muted-foreground" />
-    </div>
-  )
-}
-
-// Skeleton loading component for namespace cards
-function NamespaceCardSkeleton() {
-  return (
-    <div className="flex items-center gap-4 p-4 rounded-lg bg-secondary/30 border border-transparent animate-pulse">
-      {/* Icon placeholder */}
-      <div className="w-10 h-10 rounded-lg bg-secondary/50" />
-
-      {/* Content placeholder */}
-      <div className="flex-1 space-y-2">
-        <div className="flex items-center gap-2">
-          <div className="h-4 w-32 bg-secondary/50 rounded" />
-          <div className="h-4 w-14 bg-secondary/50 rounded" />
-        </div>
-        <div className="h-3 w-24 bg-secondary/50 rounded" />
-      </div>
-
-      {/* Cluster badge placeholder */}
-      <div className="h-6 w-20 bg-secondary/50 rounded-full" />
-
-      {/* Chevron placeholder */}
-      <div className="w-4 h-4 bg-secondary/50 rounded" />
-    </div>
-  )
-}
-
-// Common users/groups for namespace access
-const AVAILABLE_USERS = [
-  'admin@example.com',
-  'developer@example.com',
-  'operator@example.com',
-  'viewer@example.com',
-  'ci-bot@example.com',
-]
-
-const AVAILABLE_GROUPS = [
-  'developers',
-  'operators',
-  'viewers',
-  'platform-team',
-  'sre-team',
-]
-
-interface DeleteConfirmModalProps {
-  namespace: NamespaceDetails
-  onClose: () => void
-  onConfirm: () => void
-}
-
-function DeleteConfirmModal({ namespace, onClose, onConfirm }: DeleteConfirmModalProps) {
-  const [confirmText, setConfirmText] = useState('')
-  const [deleting, setDeleting] = useState(false)
-
-  const canDelete = confirmText === namespace.name
-
-  const handleDelete = async () => {
-    if (!canDelete) return
-    setDeleting(true)
-    await onConfirm()
-  }
-
-  return (
-    <BaseModal isOpen={true} onClose={onClose} size="md">
-      <BaseModal.Header
-        title="Delete Namespace"
-        description="This action cannot be undone"
-        icon={Trash2}
-        onClose={onClose}
-        showBack={false}
-      />
-
-      <BaseModal.Content>
-        <div className="mb-4 p-3 rounded-lg bg-red-500/10 border border-red-500/30">
-          <p className="text-sm text-red-300">
-            You are about to delete namespace <strong>"{namespace.name}"</strong> from cluster <strong>"{namespace.cluster}"</strong>.
-            This will permanently delete all resources within the namespace.
-          </p>
-        </div>
-
-        <div>
-          <label className="block text-sm font-medium text-muted-foreground mb-1">
-            Type <span className="text-red-400 font-mono">{namespace.name}</span> to confirm
-          </label>
-          <input
-            type="text"
-            value={confirmText}
-            onChange={(e) => setConfirmText(e.target.value)}
-            placeholder="Enter namespace name"
-            className="w-full px-3 py-2 rounded-lg bg-secondary border border-border text-white placeholder:text-muted-foreground focus:outline-none focus:ring-2 focus:ring-red-500/50"
-          />
-        </div>
-      </BaseModal.Content>
-
-      <BaseModal.Footer>
-        <div className="flex-1" />
-        <div className="flex gap-3">
-          <Button
-            variant="ghost"
-            size="lg"
-            onClick={onClose}
-          >
-            Cancel
-          </Button>
-          <Button
-            variant="danger"
-            size="lg"
-            onClick={handleDelete}
-            disabled={!canDelete || deleting}
-          >
-            {deleting ? 'Deleting...' : 'Delete Namespace'}
-          </Button>
-        </div>
-      </BaseModal.Footer>
-    </BaseModal>
-  )
-}
-
-interface CreateNamespaceModalProps {
-  clusters: string[]
-  onClose: () => void
-  onCreated: (cluster: string) => void
-}
-
-interface InitialAccessEntry {
-  type: 'User' | 'Group'
-  name: string
-  role: 'cluster-admin' | 'admin' | 'edit' | 'view'
-}
-
-function CreateNamespaceModal({ clusters, onClose, onCreated }: CreateNamespaceModalProps) {
-  const { t } = useTranslation()
-  const [name, setName] = useState('')
-  const [cluster, setCluster] = useState(clusters[0] || '')
-  const [teamLabel, setTeamLabel] = useState('')
-  const [creating, setCreating] = useState(false)
-  const [error, setError] = useState<string | null>(null)
-  const [initialAccess, setInitialAccess] = useState<InitialAccessEntry[]>([])
-  const [showUserDropdown, setShowUserDropdown] = useState(false)
-  const [showGroupDropdown, setShowGroupDropdown] = useState(false)
-
-  const addUserAccess = (user: string) => {
-    if (!initialAccess.some(a => a.type === 'User' && a.name === user)) {
-      setInitialAccess([...initialAccess, { type: 'User', name: user, role: 'edit' }])
-    }
-    setShowUserDropdown(false)
-  }
-
-  const addGroupAccess = (group: string) => {
-    if (!initialAccess.some(a => a.type === 'Group' && a.name === group)) {
-      setInitialAccess([...initialAccess, { type: 'Group', name: group, role: 'edit' }])
-    }
-    setShowGroupDropdown(false)
-  }
-
-  const removeAccess = (index: number) => {
-    setInitialAccess(initialAccess.filter((_, i) => i !== index))
-  }
-
-  const updateAccessRole = (index: number, role: 'cluster-admin' | 'admin' | 'edit' | 'view') => {
-    setInitialAccess(initialAccess.map((a, i) => i === index ? { ...a, role } : a))
-  }
-
-  const handleCreate = async () => {
-    if (!name || !cluster) return
-
-    setCreating(true)
-    setError(null)
-
-    try {
-      const labels: Record<string, string> = {}
-      if (teamLabel) {
-        labels['team'] = teamLabel
-      }
-
-      await api.post('/api/namespaces', {
-        cluster,
-        name,
-        labels: Object.keys(labels).length > 0 ? labels : undefined,
-        initialAccess: initialAccess.length > 0 ? initialAccess : undefined,
-      })
-      onCreated(cluster)
-    } catch (err: unknown) {
-      const errorMessage = err instanceof Error ? err.message : 'Failed to create namespace'
-      setError(errorMessage)
-    } finally {
-      setCreating(false)
-    }
-  }
-
-  const availableUsers = AVAILABLE_USERS.filter(
-    u => !initialAccess.some(a => a.type === 'User' && a.name === u)
-  )
-
-  const availableGroups = AVAILABLE_GROUPS.filter(
-    g => !initialAccess.some(a => a.type === 'Group' && a.name === g)
-  )
-
-  const handleClose = () => {
-    if ((name.trim() !== '' || teamLabel.trim() !== '') && !window.confirm(t('common:common.discardUnsavedChanges', 'Discard unsaved changes?'))) {
-      return
-    }
-    onClose()
-  }
-
-  return (
-    <BaseModal isOpen={true} onClose={handleClose} size="lg" closeOnBackdrop={false} closeOnEscape={true}>
-      <BaseModal.Header
-        title="Create Namespace"
-        icon={Folder}
-        onClose={handleClose}
-        showBack={false}
-      />
-
-      <BaseModal.Content className="max-h-[60vh]">
-        {error && (
-          <div className="mb-4 p-3 rounded-lg bg-red-500/20 border border-red-500/50 text-red-400 text-sm">
-            {error}
-          </div>
-        )}
-
-        <div className="space-y-4">
-          <div>
-            <label className="block text-sm font-medium text-muted-foreground mb-1">{t('common.cluster')}</label>
-            <select
-              value={cluster}
-              onChange={(e) => setCluster(e.target.value)}
-              className="w-full px-3 py-2 rounded-lg bg-secondary border border-border text-white focus:outline-none focus:ring-2 focus:ring-blue-500/50"
-            >
-              {clusters.map(c => (
-                <option key={c} value={c}>{c}</option>
-              ))}
-            </select>
-          </div>
-
-          <div>
-            <label className="block text-sm font-medium text-muted-foreground mb-1">Namespace Name</label>
-            <input
-              type="text"
-              value={name}
-              onChange={(e) => setName(e.target.value.toLowerCase().replace(/[^a-z0-9-]/g, '-'))}
-              placeholder="my-namespace"
-              className="w-full px-3 py-2 rounded-lg bg-secondary border border-border text-white placeholder:text-muted-foreground focus:outline-none focus:ring-2 focus:ring-blue-500/50"
-            />
-            <p className="text-xs text-muted-foreground mt-1">
-              Lowercase letters, numbers, and hyphens only
-            </p>
-          </div>
-
-          <div>
-            <label className="block text-sm font-medium text-muted-foreground mb-1">Team Label (optional)</label>
-            <input
-              type="text"
-              value={teamLabel}
-              onChange={(e) => setTeamLabel(e.target.value)}
-              placeholder="platform-team"
-              className="w-full px-3 py-2 rounded-lg bg-secondary border border-border text-white placeholder:text-muted-foreground focus:outline-none focus:ring-2 focus:ring-blue-500/50"
-            />
-          </div>
-
-          {/* Initial Access Section */}
-          <div>
-            <label className="block text-sm font-medium text-muted-foreground mb-2">
-              Grant Initial Access (optional)
-            </label>
-
-            {/* Add User/Group buttons */}
-            <div className="flex gap-2 mb-3">
-              <div className="relative">
-                <button
-                  onClick={() => setShowUserDropdown(!showUserDropdown)}
-                  className="flex items-center gap-1 px-3 py-1.5 rounded-lg bg-blue-500/20 text-blue-400 hover:bg-blue-500/30 transition-colors text-sm"
-                >
-                  <UserPlus className="w-4 h-4" />
-                  Add User
-                </button>
-                {showUserDropdown && availableUsers.length > 0 && (
-                  <div
-                    role="listbox"
-                    aria-label="Select user"
-                    className="absolute z-10 top-full left-0 mt-1 w-48 bg-card border border-border rounded-lg shadow-lg max-h-48 overflow-y-auto"
-                    onKeyDown={(e) => {
-                      if (e.key !== 'ArrowDown' && e.key !== 'ArrowUp') return
-                      e.preventDefault()
-                      const items = e.currentTarget.querySelectorAll<HTMLElement>('[role="option"]')
-                      const idx = Array.from(items).indexOf(document.activeElement as HTMLElement)
-                      if (e.key === 'ArrowDown') items[Math.min(idx + 1, items.length - 1)]?.focus()
-                      else items[Math.max(idx - 1, 0)]?.focus()
-                    }}
-                  >
-                    {availableUsers.map(user => (
-                      <button
-                        key={user}
-                        role="option"
-                        aria-selected={false}
-                        onClick={() => addUserAccess(user)}
-                        className="w-full px-3 py-2 text-left text-sm text-white hover:bg-secondary/50 transition-colors"
-                      >
-                        {user}
-                      </button>
-                    ))}
-                  </div>
-                )}
-              </div>
-              <div className="relative">
-                <Button
-                  variant="accent"
-                  onClick={() => setShowGroupDropdown(!showGroupDropdown)}
-                  icon={<Shield className="w-4 h-4" />}
-                >
-                  Add Group
-                </Button>
-                {showGroupDropdown && availableGroups.length > 0 && (
-                  <div
-                    role="listbox"
-                    aria-label="Select group"
-                    className="absolute z-10 top-full left-0 mt-1 w-48 bg-card border border-border rounded-lg shadow-lg max-h-48 overflow-y-auto"
-                    onKeyDown={(e) => {
-                      if (e.key !== 'ArrowDown' && e.key !== 'ArrowUp') return
-                      e.preventDefault()
-                      const items = e.currentTarget.querySelectorAll<HTMLElement>('[role="option"]')
-                      const idx = Array.from(items).indexOf(document.activeElement as HTMLElement)
-                      if (e.key === 'ArrowDown') items[Math.min(idx + 1, items.length - 1)]?.focus()
-                      else items[Math.max(idx - 1, 0)]?.focus()
-                    }}
-                  >
-                    {availableGroups.map(group => (
-                      <button
-                        key={group}
-                        role="option"
-                        aria-selected={false}
-                        onClick={() => addGroupAccess(group)}
-                        className="w-full px-3 py-2 text-left text-sm text-white hover:bg-secondary/50 transition-colors"
-                      >
-                        {group}
-                      </button>
-                    ))}
-                  </div>
-                )}
-              </div>
-            </div>
-
-            {/* Close dropdowns overlay */}
-            {(showUserDropdown || showGroupDropdown) && (
-              <button
-                onClick={() => {
-                  setShowUserDropdown(false)
-                  setShowGroupDropdown(false)
-                }}
-                className="fixed inset-0 z-0"
-                aria-label="Close dropdown"
-              />
-            )}
-
-            {/* Selected access list */}
-            {initialAccess.length > 0 && (
-              <div className="space-y-2">
-                {initialAccess.map((entry, index) => (
-                  <div
-                    key={`${entry.type}-${entry.name}`}
-                    className="flex items-center justify-between p-2 rounded-lg bg-secondary/30"
-                  >
-                    <div className="flex items-center gap-2">
-                      <span className={`text-xs px-1.5 py-0.5 rounded ${
-                        entry.type === 'User' ? 'bg-blue-500/20 text-blue-400' : 'bg-purple-500/20 text-purple-400'
-                      }`}>
-                        {entry.type}
-                      </span>
-                      <span className="text-sm text-white">{entry.name}</span>
-                    </div>
-                    <div className="flex items-center gap-2">
-                      <select
-                        value={entry.role}
-                        onChange={(e) => updateAccessRole(index, e.target.value as 'cluster-admin' | 'admin' | 'edit' | 'view')}
-                        className="px-2 py-1 text-xs rounded bg-secondary border border-border text-white"
-                      >
-                        <option value="cluster-admin">{t('namespaces.roleFullAdmin')}</option>
-                        <option value="admin">{t('namespaces.roleAdminShort')}</option>
-                        <option value="edit">{t('common.edit')}</option>
-                        <option value="view">{t('common.view')}</option>
-                      </select>
-                      <button
-                        onClick={() => removeAccess(index)}
-                        className="p-1 rounded text-muted-foreground hover:text-red-400 hover:bg-red-500/10"
-                      >
-                        <X className="w-4 h-4" />
-                      </button>
-                    </div>
-                  </div>
-                ))}
-              </div>
-            )}
-
-            {initialAccess.length === 0 && (
-              <p className="text-xs text-muted-foreground">
-                No initial access configured. You can add users/groups after creation.
-              </p>
-            )}
-          </div>
-        </div>
-      </BaseModal.Content>
-
-      <BaseModal.Footer>
-        <div className="flex-1" />
-        <div className="flex gap-3">
-          <Button
-            variant="ghost"
-            size="lg"
-            onClick={handleClose}
-          >
-            Cancel
-          </Button>
-          <Button
-            variant="primary"
-            size="lg"
-            onClick={handleCreate}
-            disabled={!name || !cluster || creating}
-          >
-            {creating ? 'Creating...' : 'Create'}
-          </Button>
-        </div>
-      </BaseModal.Footer>
-    </BaseModal>
-  )
-}
-
-interface GrantAccessModalProps {
-  namespace: NamespaceDetails
-  existingAccess: NamespaceAccessEntry[]
-  onClose: () => void
-  onGranted: () => void
-}
-
-// Common users/groups discovered from RBAC (these would ideally come from cluster role bindings)
-const COMMON_SUBJECTS = {
-  User: [
-    'admin@example.com',
-    'developer@example.com',
-    'operator@example.com',
-    'viewer@example.com',
-    'ci-bot@example.com',
-  ],
-  Group: [
-    'system:authenticated',
-    'system:cluster-admins',
-    'developers',
-    'operators',
-    'viewers',
-    'platform-team',
-    'sre-team',
-  ],
-  ServiceAccount: [
-    'default',
-    'deployer',
-    'argocd-application-controller',
-    'flux-reconciler',
-    'prometheus',
-  ],
-}
-
-function GrantAccessModal({ namespace, existingAccess, onClose, onGranted }: GrantAccessModalProps) {
-  const { t } = useTranslation()
-  const [subjectKind, setSubjectKind] = useState<'User' | 'Group' | 'ServiceAccount'>('User')
-  const [subjectName, setSubjectName] = useState('')
-  const [subjectNS, setSubjectNS] = useState('')
-  const [role, setRole] = useState('admin')
-  const [granting, setGranting] = useState(false)
-  const [error, setError] = useState<string | null>(null)
-  const [showDropdown, setShowDropdown] = useState(false)
-
-  // Filter out subjects that already have access
-  const existingSubjectNames = new Set(
-    existingAccess
-      .filter(e => e.subjectKind === subjectKind)
-      .map(e => e.subjectName)
-  )
-
-  const availableSubjects = COMMON_SUBJECTS[subjectKind].filter(
-    name => !existingSubjectNames.has(name)
-  )
-
-  const handleGrant = async () => {
-    if (!subjectName) return
-
-    setGranting(true)
-    setError(null)
-
-    try {
-      await api.post(`/api/namespaces/${namespace.name}/access`, {
-        cluster: namespace.cluster,
-        subjectKind,
-        subjectName,
-        subjectNamespace: subjectKind === 'ServiceAccount' ? subjectNS : undefined,
-        role,
-      })
-      onGranted()
-    } catch (err: unknown) {
-      const errorMessage = err instanceof Error ? err.message : 'Failed to grant access'
-      setError(errorMessage)
-    } finally {
-      setGranting(false)
-    }
-  }
-
-  const selectSubject = (name: string) => {
-    setSubjectName(name)
-    setShowDropdown(false)
-  }
-
-  const handleClose = () => {
-    if (subjectName.trim() !== '' && !window.confirm(t('common:common.discardUnsavedChanges', 'Discard unsaved changes?'))) {
-      return
-    }
-    onClose()
-  }
-
-  return (
-    <BaseModal isOpen={true} onClose={handleClose} size="md" closeOnBackdrop={false} closeOnEscape={true}>
-      <BaseModal.Header
-        title="Grant Access"
-        description={`Namespace: ${namespace.name}`}
-        icon={Shield}
-        onClose={handleClose}
-        showBack={false}
-      />
-
-      <BaseModal.Content>
-        {error && (
-          <div className="mb-4 p-3 rounded-lg bg-red-500/20 border border-red-500/50 text-red-400 text-sm">
-            {error}
-          </div>
-        )}
-
-        <div className="space-y-4">
-          <div>
-            <label className="block text-sm font-medium text-muted-foreground mb-1">Subject Type</label>
-            <select
-              value={subjectKind}
-              onChange={(e) => {
-                setSubjectKind(e.target.value as 'User' | 'Group' | 'ServiceAccount')
-                setSubjectName('') // Clear selection when type changes
-              }}
-              className="w-full px-3 py-2 rounded-lg bg-secondary border border-border text-white focus:outline-none focus:ring-2 focus:ring-blue-500/50"
-            >
-              <option value="User">{t('namespaces.subjectUser')}</option>
-              <option value="Group">{t('namespaces.subjectGroup')}</option>
-              <option value="ServiceAccount">{t('namespaces.subjectServiceAccount')}</option>
-            </select>
-          </div>
-
-          <div className="relative">
-            <label className="block text-sm font-medium text-muted-foreground mb-1">
-              {subjectKind === 'User' ? 'Username / Email' : subjectKind === 'Group' ? 'Group Name' : 'Service Account Name'}
-            </label>
-            <div className="relative">
-              <input
-                type="text"
-                value={subjectName}
-                onChange={(e) => setSubjectName(e.target.value)}
-                onFocus={() => setShowDropdown(true)}
-                placeholder={subjectKind === 'User' ? 'Select or type a user...' : subjectKind === 'Group' ? 'Select or type a group...' : 'Select or type a service account...'}
-                className="w-full px-3 py-2 rounded-lg bg-secondary border border-border text-white placeholder:text-muted-foreground focus:outline-none focus:ring-2 focus:ring-blue-500/50"
-              />
-              {showDropdown && availableSubjects.length > 0 && (
-                <div className="absolute z-10 top-full left-0 right-0 mt-1 bg-card border border-border rounded-lg shadow-lg max-h-48 overflow-y-auto">
-                  {availableSubjects
-                    .filter(name => !subjectName || name.toLowerCase().includes(subjectName.toLowerCase()))
-                    .map(name => (
-                      <button
-                        key={name}
-                        onClick={() => selectSubject(name)}
-                        className="w-full px-3 py-2 text-left text-sm text-white hover:bg-secondary/50 transition-colors"
-                      >
-                        {name}
-                      </button>
-                    ))}
-                  {subjectName && !availableSubjects.some(n => n.toLowerCase() === subjectName.toLowerCase()) && (
-                    <button
-                      onClick={() => selectSubject(subjectName)}
-                      className="w-full px-3 py-2 text-left text-sm text-blue-400 hover:bg-secondary/50 transition-colors border-t border-border"
-                    >
-                      Use "{subjectName}"
-                    </button>
-                  )}
-                </div>
-              )}
-            </div>
-            {showDropdown && (
-              <button
-                onClick={() => setShowDropdown(false)}
-                className="fixed inset-0 z-0"
-                aria-label="Close dropdown"
-              />
-            )}
-          </div>
-
-          {subjectKind === 'ServiceAccount' && (
-            <div>
-              <label className="block text-sm font-medium text-muted-foreground mb-1">Service Account Namespace</label>
-              <input
-                type="text"
-                value={subjectNS}
-                onChange={(e) => setSubjectNS(e.target.value)}
-                placeholder="default"
-                className="w-full px-3 py-2 rounded-lg bg-secondary border border-border text-white placeholder:text-muted-foreground focus:outline-none focus:ring-2 focus:ring-blue-500/50"
-              />
-            </div>
-          )}
-
-          <div>
-            <label className="block text-sm font-medium text-muted-foreground mb-1">{t('common.role')}</label>
-            <select
-              value={role}
-              onChange={(e) => setRole(e.target.value)}
-              className="w-full px-3 py-2 rounded-lg bg-secondary border border-border text-white focus:outline-none focus:ring-2 focus:ring-blue-500/50"
-            >
-              <option value="cluster-admin">{t('namespaces.roleClusterAdmin')}</option>
-              <option value="admin">{t('namespaces.roleAdmin')}</option>
-              <option value="edit">{t('namespaces.roleEdit')}</option>
-              <option value="view">{t('namespaces.roleView')}</option>
-            </select>
-            <p className="text-xs text-muted-foreground mt-1">
-              {t('namespaces.rolesNamespaceScoped')}
-            </p>
-          </div>
-        </div>
-      </BaseModal.Content>
-
-      <BaseModal.Footer>
-        <div className="flex-1" />
-        <div className="flex gap-3">
-          <Button
-            variant="ghost"
-            size="lg"
-            onClick={handleClose}
-          >
-            Cancel
-          </Button>
-          <Button
-            variant="primary"
-            size="lg"
-            onClick={handleGrant}
-            disabled={!subjectName || granting}
-          >
-            {granting ? 'Granting...' : 'Grant Access'}
-          </Button>
-        </div>
-      </BaseModal.Footer>
-    </BaseModal>
-  )
-}
