+++ web/src/components/namespaces/types.ts
@@ -1,0 +1,16 @@
+export interface NamespaceDetails {
+  name: string
+  cluster: string
+  status: string
+  labels?: Record<string, string>
+  createdAt: string
+}
+
+export interface NamespaceAccessEntry {
+  bindingName: string
+  subjectKind: string
+  subjectName: string
+  subjectNamespace?: string
+  roleName: string
+  roleKind: string
+}
